requirements.txt
boto3==1.18.14
Django==3.2.5
django-crispy-forms==1.12.0
django-storages==1.11.1
mixer==7.1.2
python-decouple==3.4
psycopg2-binary==2.9.1
pytest==6.2.4
pytest-django==4.4.0
pytest-cov==2.12.1
selenium==3.141.0
